anomaly-detector/requirements.txt
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.2
prometheus-client==0.20.0
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.49.0
kubernetes==29.0.0
python-dotenv==1.0.1
langfuse==4.5.0
